Move over last integration test
