+++ tests/spinup_tests.rs
@@ -157,7 +157,7 @@ mod spinup_tests {
                         200,
                         [
                             ("env_some_key".to_owned(), "some_value".to_owned()),
-                            ("ENV_foo".to_owned(), "bar".to_owned()),
+                            ("env_foo".to_owned(), "bar".to_owned()),
                         ]
                         .into_iter()
                         .collect(),
@@ -1050,6 +1050,72 @@ route = "/..."
                 assert_spin_request(env.runtime_mut(), request, expected)
             },
         )?;
+
+        Ok(())
+    }
+
+    // TODO: consider removing this test or rewriting
+    // We are treating the timer trigger both as an example and as a test which makes maintenance hard.
+    #[tokio::test]
+    async fn test_timer_trigger() -> anyhow::Result<()> {
+        use std::fs;
+        use std::path::Path;
+
+        const TIMER_TRIGGER_INTEGRATION_TEST: &str = "examples/spin-timer/app-example";
+        const TIMER_TRIGGER_DIRECTORY: &str = "examples/spin-timer";
+
+        let trigger_dir = Path::new(TIMER_TRIGGER_DIRECTORY);
+
+        // Conventionally, we would do all Cargo builds of test code in build.rs, but this one can take a lot
+        // longer than the tiny tests we normally build there (and it's pointless if the user just wants to build
+        // Spin without running any tests) so we do it here instead.  Subsequent builds after the first one should
+        // be very fast.
+        assert!(std::process::Command::new("cargo")
+            .arg("build")
+            // .arg("--release")
+            .arg("--target-dir")
+            .arg(trigger_dir.join("target"))
+            .arg("--manifest-path")
+            .arg(trigger_dir.join("Cargo.toml"))
+            .status()?
+            .success());
+
+        // Create a test plugin store so we don't modify the user's real one.
+        let plugin_store_dir = Path::new(concat!(env!("OUT_DIR"), "/plugin-store"));
+        let plugins_dir = plugin_store_dir.join("spin/plugins");
+
+        let plugin_dir = plugins_dir.join("trigger-timer");
+        fs::create_dir_all(&plugin_dir)?;
+        fs::copy(
+            trigger_dir.join("target/release/trigger-timer"),
+            plugin_dir.join("trigger-timer"),
+        )
+        .context("could not copy plugin binary into plugin directory")?;
+
+        let manifests_dir = plugins_dir.join("manifests");
+        fs::create_dir_all(&manifests_dir)?;
+        // Note that the hash and path in the manifest aren't accurate, but they won't be used anyway for this
+        // test. We just need something that parses without throwing errors here.
+        fs::copy(
+            Path::new(TIMER_TRIGGER_DIRECTORY).join("trigger-timer.json"),
+            manifests_dir.join("trigger-timer.json"),
+        )
+        .context("could not copy plugin manifest into manifests directory")?;
+
+        let out = std::process::Command::new(spin_binary())
+            .args([
+                "up",
+                "--file",
+                &format!("{TIMER_TRIGGER_INTEGRATION_TEST}/spin.toml"),
+                "--test",
+            ])
+            .env("TEST_PLUGINS_DIRECTORY", plugin_store_dir)
+            .output()?;
+        assert!(
+            out.status.success(),
+            "Running `spin up` returned error: {}",
+            String::from_utf8_lossy(&out.stderr)
+        );
 
         Ok(())
     }
